Changes for infra & hr modules. Cleaned up unnecessary structures. Migration needed.
