+++ facilities/models/__init__.py
@@ -1,2 +1,2 @@
 from .facility_models import *  # noqa
-from .infrastructure import *
+# from .infrastructure import *
