+++ facilities/views/__init__.py
@@ -3,4 +3,4 @@ from .facility_dashboard import *  # noqa
 from .facility_setup_views import *  # noqa
 from .facility_special_workflow_views import *  # noqa
 from .service_catalog_views import *  # noqa
-from .facility_human_resource_views import *
+# from .facility_human_resource_views import *
